Prepare to use FtOptions in energyplus::ForwardTranslator:  Break ForwardTranslatorOptions and RunOptions in two separate set of files
