+++ src/utilities/filetypes/RunOptions.cpp
@@ -29,6 +29,7 @@
 
 #include "RunOptions.hpp"
 #include "RunOptions_Impl.hpp"
+#include "ForwardTranslatorOptions.hpp"
 
 #include "../core/Assert.hpp"
 
@@ -39,97 +40,6 @@
 
 namespace openstudio {
 namespace detail {
-
-  void ForwardTranslatorOptions_Impl::reset() {
-    m_runcontrolspecialdays = true;
-    m_ip_tabular_output = false;
-    m_no_lifecyclecosts = false;
-    m_no_sqlite_output = false;
-    m_no_html_output = false;
-    m_no_variable_dictionary = false;
-    m_no_space_translation = false;
-  }
-
-  bool ForwardTranslatorOptions_Impl::keepRunControlSpecialDays() const {
-    return m_runcontrolspecialdays;
-  }
-
-  void ForwardTranslatorOptions_Impl::setKeepRunControlSpecialDays(bool keepRunControlSpecialDays) {
-    m_runcontrolspecialdays = keepRunControlSpecialDays;
-  }
-
-  bool ForwardTranslatorOptions_Impl::iPTabularOutput() const {
-    return m_ip_tabular_output;
-  }
-
-  void ForwardTranslatorOptions_Impl::setIPTabularOutput(bool iPTabularOutput) {
-    m_ip_tabular_output = iPTabularOutput;
-  }
-
-  bool ForwardTranslatorOptions_Impl::excludeLCCObjects() const {
-    return m_no_lifecyclecosts;
-  }
-
-  void ForwardTranslatorOptions_Impl::setExcludeLCCObjects(bool excludeLCCObjects) {
-    m_no_lifecyclecosts = excludeLCCObjects;
-  }
-
-  bool ForwardTranslatorOptions_Impl::excludeSQliteOutputReport() const {
-    return m_no_sqlite_output;
-  }
-
-  void ForwardTranslatorOptions_Impl::setExcludeSQliteOutputReport(bool excludeSQliteOutputReport) {
-    m_no_sqlite_output = excludeSQliteOutputReport;
-  }
-
-  bool ForwardTranslatorOptions_Impl::excludeHTMLOutputReport() const {
-    return m_no_html_output;
-  }
-
-  void ForwardTranslatorOptions_Impl::setExcludeHTMLOutputReport(bool excludeHTMLOutputReport) {
-    m_no_html_output = excludeHTMLOutputReport;
-  }
-
-  bool ForwardTranslatorOptions_Impl::excludeVariableDictionary() const {
-    return m_no_variable_dictionary;
-  }
-
-  void ForwardTranslatorOptions_Impl::setExcludeVariableDictionary(bool excludeVariableDictionary) {
-    m_no_variable_dictionary = excludeVariableDictionary;
-  }
-
-  bool ForwardTranslatorOptions_Impl::excludeSpaceTranslation() const {
-    return m_no_space_translation;
-  }
-
-  void ForwardTranslatorOptions_Impl::setExcludeSpaceTranslation(bool excludeSpaceTranslation) {
-    m_no_space_translation = excludeSpaceTranslation;
-  }
-
-  Json::Value ForwardTranslatorOptions_Impl::json() const {
-    Json::Value value;
-
-    value["runcontrolspecialdays"] = m_runcontrolspecialdays;
-    value["ip_tabular_output"] = m_ip_tabular_output;
-    value["no_lifecyclecosts"] = m_no_lifecyclecosts;
-    value["no_sqlite_output"] = m_no_sqlite_output;
-    value["no_html_output"] = m_no_html_output;
-    value["no_variable_dictionary"] = m_no_variable_dictionary;
-    value["no_space_translation"] = m_no_space_translation;
-
-    return value;
-  }
-
-  std::string ForwardTranslatorOptions_Impl::string() const {
-
-    // Write to string
-    Json::StreamWriterBuilder wbuilder;
-    // mimic the old StyledWriter behavior:
-    wbuilder["indentation"] = "   ";
-    std::string result = Json::writeString(wbuilder, json());
-
-    return result;
-  }
 
   void RunOptions_Impl::onUpdate() {
     this->onChange.nano_emit();
@@ -343,122 +253,6 @@ std::string CustomOutputAdapter::className() const {
 
 std::string CustomOutputAdapter::options() const {
   return m_options;
-}
-
-ForwardTranslatorOptions::ForwardTranslatorOptions(std::shared_ptr<detail::ForwardTranslatorOptions_Impl> impl) : m_impl(std::move(impl)) {}
-ForwardTranslatorOptions::ForwardTranslatorOptions() : m_impl(std::make_shared<detail::ForwardTranslatorOptions_Impl>()) {}
-
-void ForwardTranslatorOptions::reset() {
-  m_impl->reset();
-}
-
-boost::optional<ForwardTranslatorOptions> ForwardTranslatorOptions::fromString(const std::string& s) {
-
-  // We let it fail with a warning message
-  Json::CharReaderBuilder rbuilder;
-  std::istringstream ss(s);
-  std::string formattedErrors;
-  Json::Value value;
-  bool parsingSuccessful = Json::parseFromStream(rbuilder, ss, &value, &formattedErrors);
-  if (!parsingSuccessful) {
-    LOG(Warn, "Couldn't parse ForwardTranslatorOptions from string s='" << s << "'. Error: '" << formattedErrors << "'.");
-    return boost::none;
-  }
-
-  return ForwardTranslatorOptions::fromJSON(value);
-}
-
-ForwardTranslatorOptions ForwardTranslatorOptions::fromJSON(const Json::Value& value) {
-
-  ForwardTranslatorOptions result;
-
-  if (value.isMember("runcontrolspecialdays") && value["runcontrolspecialdays"].isBool()) {
-    result.setKeepRunControlSpecialDays(value["runcontrolspecialdays"].asBool());
-  }
-  if (value.isMember("ip_tabular_output") && value["ip_tabular_output"].isBool()) {
-    result.setIPTabularOutput(value["ip_tabular_output"].asBool());
-  }
-  if (value.isMember("no_lifecyclecosts") && value["no_lifecyclecosts"].isBool()) {
-    result.setExcludeLCCObjects(value["no_lifecyclecosts"].asBool());
-  }
-  if (value.isMember("no_sqlite_output") && value["no_sqlite_output"].isBool()) {
-    result.setExcludeSQliteOutputReport(value["no_sqlite_output"].asBool());
-  }
-  if (value.isMember("no_html_output") && value["no_html_output"].isBool()) {
-    result.setExcludeHTMLOutputReport(value["no_html_output"].asBool());
-  }
-  if (value.isMember("no_variable_dictionary") && value["no_variable_dictionary"].isBool()) {
-    result.setExcludeVariableDictionary(value["no_variable_dictionary"].asBool());
-  }
-  if (value.isMember("no_space_translation") && value["no_space_translation"].isBool()) {
-    result.setExcludeSpaceTranslation(value["no_space_translation"].asBool());
-  }
-
-  return result;
-}
-
-Json::Value ForwardTranslatorOptions::json() const {
-  return m_impl->json();
-}
-
-std::string ForwardTranslatorOptions::string() const {
-  return m_impl->string();
-}
-
-bool ForwardTranslatorOptions::keepRunControlSpecialDays() const {
-  return m_impl->keepRunControlSpecialDays();
-}
-
-void ForwardTranslatorOptions::setKeepRunControlSpecialDays(bool keepRunControlSpecialDays) {
-  m_impl->setKeepRunControlSpecialDays(keepRunControlSpecialDays);
-}
-
-bool ForwardTranslatorOptions::iPTabularOutput() const {
-  return m_impl->iPTabularOutput();
-}
-
-void ForwardTranslatorOptions::setIPTabularOutput(bool iPTabularOutput) {
-  m_impl->setIPTabularOutput(iPTabularOutput);
-}
-
-bool ForwardTranslatorOptions::excludeLCCObjects() const {
-  return m_impl->excludeLCCObjects();
-}
-
-void ForwardTranslatorOptions::setExcludeLCCObjects(bool excludeLCCObjects) {
-  m_impl->setExcludeLCCObjects(excludeLCCObjects);
-}
-
-bool ForwardTranslatorOptions::excludeSQliteOutputReport() const {
-  return m_impl->excludeSQliteOutputReport();
-}
-
-void ForwardTranslatorOptions::setExcludeSQliteOutputReport(bool excludeSQliteOutputReport) {
-  m_impl->setExcludeSQliteOutputReport(excludeSQliteOutputReport);
-}
-
-bool ForwardTranslatorOptions::excludeHTMLOutputReport() const {
-  return m_impl->excludeHTMLOutputReport();
-}
-
-void ForwardTranslatorOptions::setExcludeHTMLOutputReport(bool excludeHTMLOutputReport) {
-  m_impl->setExcludeHTMLOutputReport(excludeHTMLOutputReport);
-}
-
-bool ForwardTranslatorOptions::excludeVariableDictionary() const {
-  return m_impl->excludeVariableDictionary();
-}
-
-void ForwardTranslatorOptions::setExcludeVariableDictionary(bool excludeVariableDictionary) {
-  m_impl->setExcludeVariableDictionary(excludeVariableDictionary);
-}
-
-bool ForwardTranslatorOptions::excludeSpaceTranslation() const {
-  return m_impl->excludeSpaceTranslation();
-}
-
-void ForwardTranslatorOptions::setExcludeSpaceTranslation(bool excludeSpaceTranslation) {
-  m_impl->setExcludeSpaceTranslation(excludeSpaceTranslation);
 }
 
 RunOptions::RunOptions() : m_impl(std::make_shared<detail::RunOptions_Impl>()) {
